pyproject.toml
[project]
name = "photojoo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115.4",
    "httpx>=0.27.2",
    "loguru>=0.7.2",
    "meilisearch>=0.31.6",
    "pillow>=11.0.0",
    "pinecone[grpc]>=5.3.1",
    "requests>=2.32.3",
    "torch>=2.5.1",
    "tqdm>=4.67.0",
    "transformers>=4.46.2",
]
